requirements.txt
python-frontmatter==1.1.0
PyYAML==6.0.2
setuptools==80.7.1
wheel==0.45.1
